moved all origanization files into Organizations directory, redirected user login to Organizations.js page/component
